requirements.txt
certifi==2020.6.20
cffi==1.14.2
cryptography==3.1
decorator==4.4.2
ephem==3.7.7.1
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.8
requests==2.7.0
six==1.15.0
telebot==0.0.3
tornado==6.0.4
emoji==0.6.0